requirements.txt
util@git+https://git@github.com/dertilo/util.git#egg=util
sklearn
pandas